Extract equals generation helpers to new file
