+++ poko-compiler-plugin/src/main/kotlin/dev/drewhamilton/poko/ir/annotationReferences.kt
@@ -1,0 +1,10 @@
+package dev.drewhamilton.poko.ir
+
+import org.jetbrains.kotlin.name.ClassId
+import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.name.Name
+
+internal val ArrayContentBasedAnnotation = ClassId(
+    FqName("dev.drewhamilton.poko"),
+    Name.identifier("ArrayContentBased"),
+)
